+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -334,7 +334,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWHSVZPCZVQPA62Z1KTW5XN0"
+        "01KWK8B2EYVVP9J95A8SRH6GEE"
     )
     r"""The unique identifier of the data source"""