+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponseragas.py
@@ -59,9 +59,9 @@ class EvaluatorResponseRagas(BaseModel):
 
     model: str
 
-    created: Optional[str] = "2026-07-02T16:16:01.733Z"
+    created: Optional[str] = "2026-07-03T05:48:17.067Z"
 
-    updated: Optional[str] = "2026-07-02T16:16:01.734Z"
+    updated: Optional[str] = "2026-07-03T05:48:17.067Z"
 
     updated_by_id: OptionalNullable[str] = UNSET