+++ packages/orq-rc/src/orq_ai_sdk/models/runagentop.py
@@ -2010,7 +2010,7 @@ class RunAgentAgentToolInputRunTools(BaseModel):
 
     schema_: Annotated[AgentToolInputRunSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KWHSVZ8T58Y4X4FB8JEWJWCT"
+    id: Optional[str] = "01KWK8B21KYWY5CZVASMTZQKJ6"
 
     description: Optional[str] = None